requirements.txt
torch
numpy
fastapi
uvicorn
python-dotenv
requests
matplotlib
scikit-learn
scrapegraphai
python-multipart
